+++ CHANGELOG.md
@@ -6,12 +6,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][v0.2.0](https://github.com/linna/csrf-guard/compare/v0.1.0...v0.2.0) - 2018-XX-XX
+
+### Added
+* `Linna\Filter\Rules\NumberCompare` validate and compare numbers with operators <, >, >=, <=, =
+
+### Removed
+* `Linna\Filter\Rules\Max`
+* `Linna\Filter\Rules\Min`
+
 ## [Initial Release][v0.1.0](https://github.com/linna/csrf-guard/compare/v0.1.0...master) - 2018-05-29
 
 ### Added
 * `Linna\Filter\Rules\Between` validate numbers, between two values
 * `Linna\Filter\Rules\Date` validate a date
-* `Linna\Filter\Rules\DateCompare` validate dates, lower than
+* `Linna\Filter\Rules\DateCompare` validate dates with operators <, >, >=, <=, =
 * `Linna\Filter\Rules\Email` validate an email
 * `Linna\Filter\Rules\Escape` escape output
 * `Linna\Filter\Rules\Max` validate numbers, lower than